Move some stuff around.
